+++ modules/services/nextcloud-server.nix
@@ -977,9 +977,9 @@ in
 
       shb.authelia.oidcClients = lib.mkIf (cfg.apps.sso.provider == "Authelia") [
         {
-          id = cfg.apps.sso.clientID;
-          description = "Nextcloud";
-          secret.source = cfg.apps.sso.secretFileForAuthelia;
+          client_id = cfg.apps.sso.clientID;
+          client_name = "Nextcloud";
+          client_secret.source = cfg.apps.sso.secretFileForAuthelia;
           public = false;
           authorization_policy = cfg.apps.sso.authorization_policy;
           redirect_uris = [ "${protocol}://${fqdnWithPort}/apps/oidc_login/oidc" ];